requirements.txt
setuptools
wheel
utils
requests
PyYAML